[layers] Collapse the quantized FC fork onto ComputeOps::fc
FullyConnectedLayerCl stops owning an OpenCL GEMM and becomes a thin
backend-neutral layer: it binds weight and bias, then dispatches the matmul
through input.getOps()->fc(input, weight, output). The eager one-time weight
transform at load time moves to the same table as fc_prebuild_weight(), a
no-op on a backend that needs none.

Because the class no longer names a backend it also stops living under one, so
it moves from layers/cl_layers/ to layers/ and is compiled unconditionally.
The header name and the "fully_connected" type string are unchanged, so
cl_context.cpp and unittest_layers_fully_connected_cl.cpp keep resolving it.

On this commit the CPU column is complete -- CpuComputeOps::fc is the host
Tensor::dot -- and the accelerator column is not: an OpenCL ComputeOps::fc
implementation is a separate change. That ordering is deliberate; the table
entry and its neutral caller have to exist before a backend can fill them in.

The layer also honours a fused activation epilogue through the op table, so a
realized conv/fc + activation pair costs no extra node here either.
