+++ nntrainer/layers/cl_layers/meson.build
@@ -1,5 +1,4 @@
 cl_layer_sources = [
-   'fc_layer_cl.cpp',
    'addition_layer_cl.cpp',
    'swiglu_cl.cpp',
    'reshape_cl.cpp',
